media-catalog/requirements.txt
guessit>=3.8.0
psycopg2-binary>=2.9
httpx>=0.27
python-dotenv>=1.0
uvicorn>=0.30
starlette>=0.38
jinja2>=3.1
click>=8.1
tqdm>=4.66
